+++ core/kernel/arch/x86_64/user_mode.h
@@ -23,6 +23,7 @@ struct x86_64_user_mode_state {
     u32 ticks_ok;
     u32 sleep_ticks_ok;
     u32 input_status_ok;
+    u32 input_events_ok;
     u32 exec_ok;
     u32 exit_ok;
     u32 unexpected_syscall;